CustomFW: Able to use printf() now
Utilizing RedLib so that we can use their implementation of printf. This is
 more robust than ConsolePrint() that was being used before (especially given
 1024 char output limit).

Also switched USB CDC UART to use a FIFO for transmitting data. This is more
 efficient than previous approach.

Added a test function for printing, but it needs more features, etc. to allow
 for more robust testing of print capabilities.

Still have at least one case of print related issues (i.e. adcRead) to look
 into.